+++ test/bulma.test.tsx
@@ -55,13 +55,10 @@ describe('removeProps functions', () => {
     });
 
     it('should remove Heading props', () => {
-        const HeadingProps = {
-            is1: true,
-            is2: true,
-            is3: true,
-            is4: true,
-            is5: true,
-            is6: true,
+        const HeadingProps: Bulma.Bulma.Heading = {
+            isHeading: 3,
+            isSize: 2,
+            isParagraph: true,
             isSpaced: true,
         }
         expect(Bulma.removeHeadingProps({ ...props, ...HeadingProps }))
@@ -152,23 +149,13 @@ describe('get*Modifiers functions', () => {
     });
 
     it('should getHeadingModifiers', () => {
-        const props = {
-            is1: true,
-            is2: true,
-            is3: true,
-            is4: true,
-            is5: true,
-            is6: true,
+        const props: any = {
+            isSize: 1,
             isSpaced: true,
         }
         const expected = {
-            'is-1': props.is1,
-            'is-2': props.is2,
-            'is-3': props.is3,
-            'is-4': props.is4,
-            'is-5': props.is5,
-            'is-6': props.is6,
-            'is-spaced': props.isSpaced,
+            'is-1': true,
+            'is-spaced': true,
         }
         expect(Bulma.getHeadingModifiers(props))
             .toEqual(expected);
@@ -177,13 +164,34 @@ describe('get*Modifiers functions', () => {
 
 describe('getHeadingElement', () => {
     it('should return h1', () => {
-        const props = {
-            is1:true,
-            is3:false,
-            is4:true,
-        }
-        const actual = Bulma.getHeadingElement(props);
-        expect(actual).toBe('h1');
+        expect(Bulma.getHeadingElement({ isHeading: 1 })).toBe('h1');
+
+        const fixture: any[] = [
+            { 1: null },
+            { isHeading: '1000' },
+            { isHeading: 1000 },
+            { isHeading: null },
+            { isHeading: undefined },
+        ];
+
+        fixture.forEach(item => expect(Bulma.getHeadingElement(item)).toBe('h1'));
+
+    });
+
+    it('should return h2', () => {
+        expect(Bulma.getHeadingElement({ isHeading: 2 })).toBe('h2');
+    });
+    it('should return h3', () => {
+        expect(Bulma.getHeadingElement({ isHeading: 3 })).toBe('h3');
+    });
+    it('should return h4', () => {
+        expect(Bulma.getHeadingElement({ isHeading: 4 })).toBe('h4');
+    });
+    it('should return h5', () => {
+        expect(Bulma.getHeadingElement({ isHeading: 5 })).toBe('h5');
+    });
+    it('should return h6', () => {
+        expect(Bulma.getHeadingElement({ isHeading: 6 })).toBe('h6');
     });
 });
 
@@ -217,14 +225,216 @@ describe('withHelpersModifiers', () => {
             )
         }
         const WithHelpersModifiersComponent = Bulma.withHelpersModifiers(Component);
-        const renderedComponent = <WithHelpersModifiersComponent isBlack isFlexDesktopOnly isFullWidth className='custom' />
+        const renderedComponent = <WithHelpersModifiersComponent isBlack isFullWidth className='custom' />
         const shallowedComponent = shallow(renderedComponent);
 
         expect(shallowedComponent.prop('isBlack')).toBe(true);
-        expect(shallowedComponent.prop('isFlexDesktopOnly')).toBe(undefined);
         expect(shallowedComponent.prop('isFullWidth')).toBe(undefined);
         expect(shallowedComponent.hasClass('custom')).toBe(true);
-        expect(shallowedComponent.hasClass('is-flex-desktop-only')).toBe(true);
         expect(shallowedComponent.hasClass('is-fullwidth')).toBe(true);
+    });
+
+    it('should render a Component with .flex', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent = Bulma.withHelpersModifiers(Component);
+        const renderedComponent = <WithHelpersModifiersComponent isFlex className='custom' />
+        const shallowedComponent = shallow(renderedComponent);
+
+        expect(shallowedComponent.prop('isFlex')).toBe(undefined);
+        expect(shallowedComponent.hasClass('is-flex')).toBe(true);
+        expect(shallowedComponent.hasClass('custom')).toBe(true);
+    });
+
+    it('should render a Component with flex modifiers', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent = Bulma.withHelpersModifiers(Component);
+        const renderedComponent = <WithHelpersModifiersComponent isFlex={['mobile',
+            'tablet',
+            'touch',
+            'desktop',
+            'widescreen']} className='custom' />
+        const shallowedComponent = shallow(renderedComponent);
+
+        expect(shallowedComponent.prop('isFlex')).toBe(undefined);
+        expect(shallowedComponent.hasClass('is-flex-mobile')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-tablet')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-touch')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-desktop')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-widescreen')).toBe(true);
+        expect(shallowedComponent.hasClass('custom')).toBe(true);
+    });
+
+    it('should render a Component without isFlex prop when someone provides bad data', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent: any = Bulma.withHelpersModifiers(Component);
+        const renderedComponent = <WithHelpersModifiersComponent isFlex='none' className='custom' />
+        const shallowedComponent = shallow(renderedComponent);
+
+        expect(shallowedComponent.prop('isFlex')).toBe(undefined);
+        expect(shallowedComponent.hasClass('custom')).toBe(true);
+    });
+
+    it('should render a Component withouth errors when passing bad data to isFlex', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent: any = Bulma.withHelpersModifiers(Component);
+        const renderedComponent = <WithHelpersModifiersComponent isFlex={1} className='custom' />
+        const shallowedComponent = shallow(renderedComponent);
+
+        expect(shallowedComponent.prop('isFlex')).toBe(undefined);
+        expect(shallowedComponent.hasClass('custom')).toBe(true);
+    });
+
+    it('should render a Component with flex "shortcut" modifiers', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent = Bulma.withHelpersModifiers(Component);
+        const renderedComponent = <WithHelpersModifiersComponent isFlex={['m', 'tt', 't', 'd', 'w']} className='custom' />
+        const shallowedComponent = shallow(renderedComponent);
+
+        expect(shallowedComponent.prop('isFlex')).toBe(undefined);
+        expect(shallowedComponent.hasClass('is-flex-mobile')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-tablet')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-touch')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-desktop')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-widescreen')).toBe(true);
+        expect(shallowedComponent.hasClass('custom')).toBe(true);
+    });
+
+    it('should render a Component with flex-only modifiers as string', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent = Bulma.withHelpersModifiers(Component);
+        const renderedComponent = <WithHelpersModifiersComponent isFlexOnly='d' className='custom' />
+        const shallowedComponent = shallow(renderedComponent);
+
+        expect(shallowedComponent.prop('isFlexOnly')).toBe(undefined);
+        expect(shallowedComponent.hasClass('is-flex-desktop-only')).toBe(true);
+        expect(shallowedComponent.hasClass('custom')).toBe(true);
+    });
+
+    it('should render a Component without isFlexOnly prop when someone provides bad data', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent: any = Bulma.withHelpersModifiers(Component);
+        const renderedComponent = <WithHelpersModifiersComponent isFlexOnly='none' className='custom' />
+        const shallowedComponent = shallow(renderedComponent);
+
+        expect(shallowedComponent.prop('isFlex')).toBe(undefined);
+        expect(shallowedComponent.hasClass('custom')).toBe(true);
+    });
+
+    it('should render a Component with flex-only modifiers as Array', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent = Bulma.withHelpersModifiers(Component);
+        const renderedComponent = <WithHelpersModifiersComponent isFlexOnly={['tablet', 'desktop']} className='custom' />
+        const shallowedComponent = shallow(renderedComponent);
+
+        expect(shallowedComponent.prop('isFlex')).toBe(undefined);
+        expect(shallowedComponent.hasClass('is-flex-tablet-only')).toBe(true);
+        expect(shallowedComponent.hasClass('is-flex-desktop-only')).toBe(true);
+        expect(shallowedComponent.hasClass('custom')).toBe(true);
+    });
+
+    it('should render a Component with hasAlignedText modifiers', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent: any = Bulma.withHelpersModifiers(Component);
+
+        expect(
+            shallow(<WithHelpersModifiersComponent hasAlignedText='l' className='custom' />)
+                .hasClass('has-text-left')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasAlignedText='c' className='custom' />)
+                .hasClass('has-text-centered')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasAlignedText='r' className='custom' />)
+                .hasClass('has-text-right')
+        ).toBe(true);
+
+        expect(
+            shallow(<WithHelpersModifiersComponent hasAlignedText='r,l' className='custom' />)
+                .hasClass('has-text-right')
+        ).toBe(false);
+    });
+
+    it('should render a Component with hasColor modifiers', () => {
+        const Component: React.SFC<React.HTMLProps<HTMLDivElement>> = (props) => {
+            return (
+                <div>Hello World</div>
+            )
+        }
+        const WithHelpersModifiersComponent = Bulma.withHelpersModifiers(Component);
+
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='white' className='custom' />)
+                .hasClass('has-text-white')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='light' className='custom' />)
+                .hasClass('has-text-light')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='dark' className='custom' />)
+                .hasClass('has-text-dark')
+        ).toBe(true);
+
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='black' className='custom' />)
+                .hasClass('has-text-black')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='primary' className='custom' />)
+                .hasClass('has-text-primary')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='info' className='custom' />)
+                .hasClass('has-text-info')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='success' className='custom' />)
+                .hasClass('has-text-success')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='warning' className='custom' />)
+                .hasClass('has-text-warning')
+        ).toBe(true);
+        expect(
+            shallow(<WithHelpersModifiersComponent hasColor='danger' className='custom' />)
+                .hasClass('has-text-danger')
+        ).toBe(true);
     });
 });